requirements.txt
great-expectations==0.18.19
pandas==2.1.4
numpy==1.26.4
sqlalchemy==2.0.36
faker==22.7.0
pytest==7.4.4
jupyter==1.0.0
python-dotenv==1.0.1
pyarrow==14.0.2
openpyxl==3.1.5
